requirements.txt
discord.py==2.3.2
together==0.2.10
python-dotenv==1.0.1
